vim-vundle added in makelinks.bash
Added function to setup vim-vundle (clone from github if it
doesn't exist.
Further improved the backup of original configuration files.

+++ makelinks.bash
@@ -6,6 +6,10 @@ dir=~/rat_dotfiles           # dotfiles directory
 olddir=~/dotfiles_old        # old dotfiles backup directory
 files="bashrc vimrc screenrc tmux.conf" # list of files/folders to symlink in homedir
 
+setup_vundle() {
+    mkdir -p .vim/bundle
+    git clone https://github.com/gmarik/vundle.git ~/.vim/bundle/vundle
+}
 
 # create dotfiles_old in homedir
 echo "Creating $olddir for backup of any existing dotfiles in ~"
@@ -19,8 +23,13 @@ echo "...done"
 
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks
 for file in $files; do
-    echo "Move $file from ~ to $olddir"
-    mv ~/.$file ~/dotfiles_old/
-    echo "Creating symlink to $file in home directory."
-    ln -s $dir/$file ~/.$file
+    if [ -f ~/.$file ]; then
+        echo "Move $file from ~ to $olddir"
+        mv ~/.$file ~/dotfiles_old/
+        echo "Creating symlink to $file in home directory."
+        ln -s $dir/$file ~/.$file
+    fi
 done
+
+[ ! -d ~/.vim/bundle/vundle ] && setup_vundle
+
